requirements.txt
python-dotenv==1.0.0
pymongo==4.5.0
flask==2.3.2
python-telegram-bot==20.3
requests==2.31.0
gunicorn==20.1.0